refactor!: move DbusActivation to its own module
